Move `FullEvent` into its own module under `model::event`
Also renames `update_cache_with_event` to `FullEvent::from_event`.
