+++ src/gateway/client/mod.rs
@@ -40,7 +40,7 @@ use tracing::instrument;
 use tracing::{debug, warn};
 
 pub use self::context::Context;
-pub use self::event_handler::{EventHandler, FullEvent, RawEventHandler};
+pub use self::event_handler::{EventHandler, RawEventHandler};
 #[cfg(feature = "voice")]
 use super::VoiceGatewayManager;
 use super::{
